@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import OdsInput from './input';
 import OdsInputVue from './input.vue';
 
@@ -6,8 +6,11 @@ document.addEventListener('DOMContentLoaded', () => {
   OdsInput.init();
 
   if (document.getElementById('ods-input-vue-1')) {
-    Vue.component('ods-input', OdsInputVue);
-    const vm = new Vue();
-    vm.$mount('#ods-input-vue-1');
+    const app = createApp({
+      components: {
+        OdsInput,
+      },
+    });
+    app.mount('#ods-input-vue-1');
   }
 });
