@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import OdsSearchInline from './inline.vue';
 
 function OdsSearchVue() {
@@ -6,7 +6,7 @@ function OdsSearchVue() {
   vueElements.forEach((element) => {
     const odsSearchPlaceholder = document.getElementById(element);
     if (odsSearchPlaceholder) {
-      const vm = new Vue({
+      const app = createApp({
         components: {
           OdsSearchInline,
         },
@@ -45,7 +45,7 @@ function OdsSearchVue() {
           },
         },
       });
-      vm.$mount(odsSearchPlaceholder);
+      app.mount(odsSearchPlaceholder);
     }
   });
 }
