@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import OdsMap from './map.vue';
 
 document.addEventListener('DOMContentLoaded', () => {
@@ -14,23 +14,26 @@ document.addEventListener('DOMContentLoaded', () => {
   const mapElement10 = document.getElementById('ods-map10');
 
   if (mapElement) {
-    Vue.component('ods-map', OdsMap);
-    const vm1 = new Vue();
-    vm1.$mount(mapElement);
+    const app1 = createApp({
+      components: { OdsMap },
+    });
+    app1.mount(mapElement);
   }
   if (mapElement2) {
-    Vue.component('ods-map', OdsMap);
-    const vm2 = new Vue();
-    vm2.$mount(mapElement2);
+    const app2 = createApp({
+      components: { OdsMap },
+    });
+    app2.mount(mapElement2);
   }
   if (mapElement3) {
-    Vue.component('ods-map', OdsMap);
-    const vm3 = new Vue();
-    vm3.$mount(mapElement3);
+    const app3 = createApp({
+      components: { OdsMap },
+    });
+    app3.mount(mapElement3);
   }
   if (mapElement4) {
-    Vue.component('ods-map', OdsMap);
-    const vm4 = new Vue({
+    const app1 = createApp({
+      components: { OdsMap },
       mounted() {
         this.$refs.odsMap4.resize();
       },
@@ -39,24 +42,26 @@ document.addEventListener('DOMContentLoaded', () => {
         resize() {},
       },
     });
-    vm4.$mount(mapElement4);
+    app4.mount(mapElement4);
   }
 
   if (mapElement5) {
-    Vue.component('ods-map', OdsMap);
-    const vm5 = new Vue();
-    vm5.$mount(mapElement5);
+    const app5 = createApp({
+      components: { OdsMap },
+    });
+    app5.mount(mapElement5);
   }
 
   if (mapElement6) {
-    Vue.component('ods-map', OdsMap);
-    const vm6 = new Vue();
-    vm6.$mount(mapElement6);
+    const app6 = createApp({
+      components: { OdsMap },
+    });
+    app6.mount(mapElement6);
   }
 
   if (mapElement7) {
-    Vue.component('ods-map', OdsMap);
-    const vm7 = new Vue({
+    const app7 = createApp({
+      components: { OdsMap },
       data: () => ({
         points: [{ longitude: 10.74981, latitude: 59.913008, popupContent: '<h2>Bob the Builder</h2>' }],
       }),
@@ -67,12 +72,12 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 10000);
       },
     });
-    vm7.$mount(mapElement7);
+    app7.mount(mapElement7);
   }
 
   if (mapElement8) {
-    Vue.component('ods-map', OdsMap);
-    const vm8 = new Vue({
+    const app8 = createApp({
+      components: { OdsMap },
       data: () => ({
         geoJson: 'https://ukeweb-public.s3.eu-central-1.amazonaws.com/map/data/kindergarten-with-events.geojson',
       }),
@@ -83,12 +88,12 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm8.$mount(mapElement8);
+    app8.mount(mapElement8);
   }
 
   if (mapElement9) {
-    Vue.component('ods-map', OdsMap);
-    const vm9 = new Vue({
+    const app9 = createApp({
+      components: { OdsMap },
       data: () => ({
         state: { longitude: 10.74981, latitude: 59.913008, zoom: 10, showPopups: true, autoFitToBounds: false },
       }),
@@ -99,17 +104,17 @@ document.addEventListener('DOMContentLoaded', () => {
         }, 15000);
       },
     });
-    vm9.$mount(mapElement9);
+    app9.mount(mapElement9);
   }
 
   if (mapElement10) {
-    Vue.component('ods-map', OdsMap);
-    const vm10 = new Vue({
+    const app10 = createApp({
+      components: { OdsMap },
       data: () => ({
         loadMap: false,
         state: { longitude: 10.74981, latitude: 59.913008, zoom: 10, showPopups: true, autoFitToBounds: false },
       }),
     });
-    vm10.$mount(mapElement10);
+    app10.mount(mapElement10);
   }
 });
